requirements.txt
streamlit==1.42.1
azure-cognitiveservices-vision-computervision==0.9.1
